911961# Refactor Language Translation Service
- Move Connector implementation (Translate and GetIdentifiableLanguages) to a new LanguageTranslationService class
- Add missing class from the previous commit, TextToSpeechService
