+++ javasource/watsonservices/actions/GetIdentifiableLanguages.java
@@ -9,17 +9,10 @@
 
 package watsonservices.actions;
 
-import java.util.ArrayList;
-import java.util.List;
-import com.ibm.watson.developer_cloud.language_translation.v2.LanguageTranslation;
-import com.ibm.watson.developer_cloud.language_translation.v2.model.IdentifiableLanguage;
-import com.mendix.core.Core;
-import com.mendix.logging.ILogNode;
-import com.mendix.systemwideinterfaces.MendixException;
 import com.mendix.systemwideinterfaces.core.IContext;
 import com.mendix.systemwideinterfaces.core.IMendixObject;
 import com.mendix.webui.CustomJavaAction;
-import watsonservices.proxies.Language;
+import watsonservices.utils.LanguageTranslationService;
 
 public class GetIdentifiableLanguages extends CustomJavaAction<java.util.List<IMendixObject>>
 {
@@ -37,30 +30,7 @@ public class GetIdentifiableLanguages extends CustomJavaAction<java.util.List<IM
 	public java.util.List<IMendixObject> executeAction() throws Exception
 	{
 		// BEGIN USER CODE
-		LOGGER.debug("Executing IdentifiableLanagues Connector...");
-
-		final LanguageTranslation service = new LanguageTranslation();
-		service.setUsernameAndPassword(username, password);
-
-	    List<IdentifiableLanguage> identifieableLanguages;
-		try{
-			identifieableLanguages = service.getIdentifiableLanguages().execute();
-		} catch (Exception e) {
-			LOGGER.error("Watson Service Connection - Failed retrieving the identifiable languages", e);
-			throw new MendixException(e);
-		}
-
-		final List<IMendixObject> results = new ArrayList<IMendixObject>();
-		for(IdentifiableLanguage language : identifieableLanguages){
-			
-			IMendixObject result = Core.instantiate(getContext(), Language.entityName);
-
-			result.setValue(getContext(), "Name", language.getName());
-			result.setValue(getContext(), "Code", language.getLanguage());
-			results.add(result);
-		}
-
-		return results;
+		return LanguageTranslationService.getIdentifiableLanguages(getContext(), username, password);
 		// END USER CODE
 	}
 
@@ -74,7 +44,5 @@ public class GetIdentifiableLanguages extends CustomJavaAction<java.util.List<IM
 	}
 
 	// BEGIN EXTRA CODE
-	private static final String WATSON_TRANSLATE_LOGNODE = "WatsonServices.IBM_WatsonConnector_Translate";
-	private static ILogNode LOGGER = Core.getLogger(Core.getConfiguration().getConstantValue(WATSON_TRANSLATE_LOGNODE).toString());
 	// END EXTRA CODE
 }
