+++ javasource/watsonservices/actions/Translate.java
@@ -9,16 +9,10 @@
 
 package watsonservices.actions;
 
-import org.apache.commons.lang3.StringUtils;
-import com.ibm.watson.developer_cloud.language_translation.v2.LanguageTranslation;
-import com.ibm.watson.developer_cloud.language_translation.v2.model.Language;
-import com.ibm.watson.developer_cloud.language_translation.v2.model.TranslationResult;
-import com.mendix.core.Core;
-import com.mendix.logging.ILogNode;
-import com.mendix.systemwideinterfaces.MendixException;
 import com.mendix.systemwideinterfaces.core.IContext;
 import com.mendix.systemwideinterfaces.core.IMendixObject;
 import com.mendix.webui.CustomJavaAction;
+import watsonservices.utils.LanguageTranslationService;
 
 /**
  * FromLang could be empty. Language will be detected if possible. If not possible an exception will be thrown
@@ -45,29 +39,7 @@ public class Translate extends CustomJavaAction<IMendixObject>
 		this.translation = __translation == null ? null : watsonservices.proxies.Translation.initialize(getContext(), __translation);
 
 		// BEGIN USER CODE
-		LOGGER.debug("Executing Translate Connector...");
-		
-		final LanguageTranslation service = new LanguageTranslation();
-		service.setUsernameAndPassword(username, password);
-
-		final Language source = getLanguage(translation.getTranslation_SourceLanguage().getCode());
-		final Language target = getLanguage(translation.getTranslation_TargetLanguage().getCode());
-
-		TranslationResult result;
-		try {
-
-			result = service.translate(translation.getText(), source, target).execute();
-		} catch (Exception e) {
-			LOGGER.error("Watson Service connection - Failed translating from" + source + " to " + target + " the text " + StringUtils.abbreviate(translation.getText(), 20), e);
-			throw new MendixException(e);
-		}
-
-		translation.setWordCount(Long.valueOf(result.getWordCount()));
-		translation.setCharacterCount(Long.valueOf(result.getCharacterCount()));
-		translation.setOutput(result.getFirstTranslation());
-		Core.commit(getContext(), translation.getMendixObject());
-
-		return translation.getMendixObject();
+		return LanguageTranslationService.translate(getContext(), translation, username, password);
 		// END USER CODE
 	}
 
@@ -81,30 +53,5 @@ public class Translate extends CustomJavaAction<IMendixObject>
 	}
 
 	// BEGIN EXTRA CODE
-	private static final String WATSON_TRANSLATE_LOGNODE = "WatsonServices.IBM_WatsonConnector_Translate";
-	private static ILogNode LOGGER = Core.getLogger(Core.getConfiguration().getConstantValue(WATSON_TRANSLATE_LOGNODE).toString());
-
-	private static Language getLanguage(String lang) throws MendixException {
-		if(Language.ARABIC.toString().equals(lang)){
-			return Language.ARABIC;
-		}
-		if(Language.ENGLISH.toString().equals(lang)){
-			return Language.ENGLISH;
-		}
-		if(Language.SPANISH.toString().equals(lang)){
-			return Language.SPANISH;
-		}
-		if(Language.FRENCH.toString().equals(lang)){
-			return Language.FRENCH;
-		}
-		if(Language.ITALIAN.toString().equals(lang)){
-			return Language.ITALIAN;
-		}
-		if(Language.PORTUGUESE.toString().equals(lang)){
-			return Language.PORTUGUESE;
-		}
-		
-		throw new MendixException("The language is not supported: " + lang);
-	}
 	// END EXTRA CODE
 }
